feat: improvements
