+++ packages/ui/src/components/molecules/SearchProducts/styles.scss
@@ -90,7 +90,7 @@
   }
 
 	[data-fs-search-product-item-content] {
-		width: 100%;
+		flex-grow: 1;
 	}
 
   [data-fs-search-product-item-image] {
@@ -123,11 +123,13 @@
     }
   }
 
-	[data-fs-search-product-item-content-wrapper] {
+	[data-fs-search-product-item-quick-order] {
 		display: flex;
+		justify-content: space-between;
+		width: 100%;
 	}
 
-	[data-fs-search-product-item-badge] {
+	[data-fs-search-product-item-quick-order-badge] {
 		margin-bottom: var(--fs-search-product-item-quick-order-spacing-0);
 
 		& + p {
